pipelines/journal-fetch/requirements.txt
feedparser>=6.0.0
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
astrapy>=1.4.0
